Added tasks in nova and neutron test yml files
Moved Rabbitmq and DB tasks to generic files.
This patch removes the redundant behaviour of the tasks such
as Ensuring Rabbitmq hosts and user, creating db, granting access
to db in the testing yml files. The patch  makes it more generic
and shifted the common tasks to separate files create-grant-db.yml
 and ensure-rabbitmq.yml

Change-Id: Ibd94e1ebdce45b782ab447ca3db5068588a1e368