Added few pre and post tasks to test-prepare-host.yml
The current version of test-prepare-host fails to check
the iptable rules which is used for lxc natting, network
configuration (checking for interfaces.d, automatic loading
of interfaces.d config files, start the network) which is used
for checking the linux bridge configuration and for installing
lxc-python2 pip package for python2 bindings.

This patch would allow test-prepare-host to perform the above
checks.

Change-Id: I0fe6c0b0f637129322ee1bc266de839b716829c4