Add convergence test for test repo.
For test repo we should build out keystone + infra services, as they are
the most commonly used roles/tasks.

We still need to add some funcitonal tests but this should give a good
base. The test repo will consume it's own test plays in the same way as
other repositories would. We use a "git clone ." to avoid a situation
where we clone the old repo containing the errors we are trying to fix.

Additional fix, we don't need a "when" clause on the rabbit/db creation
tasks - since these are already delegated and a "run_once" will suffice.
This fix is required for the tests to pass.

Change-Id: I482500872f9e29a71e1576f67a0ad390be2a3ece