Remove pinned vars and set become:true
This patch removes the following pinned vars for test execution:

lxc_container_release:
This variable is set based on the host platform and should not
be fixed to this value in the playbook.

lxc_container_backing_store:
This value is dynamically set based on the host configuration
and 'dir' is the fallback value so it does not need to be set.

global_environment_variables:
This value is an *optional* variable. It was only ever set in
order to test that the functionality to apply the contents of
this variable work.

This patch also sets 'become: true' for the playbook to ensure
that the execution of the container creation/deletion is executed
as the root user.

Change-Id: Id6f35e1b9fef5096c06dffb1900c2d144bc5ece4
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>