Fix container destroy/create
The container destroy playbook was attempting to ssh to the targets
it was destroying instead of simply connecting to the localhost and
removing them. This change allows the connection to be local allowing
the destroy to go through.

This also changes the container create and destroy tasks to be done
in serial. This is being done due to potential control path issues.

Change-Id: Ide0fa39bfef917abd8b0abf1b5c9991beae9c10b
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>