Include openstack_hosts in test-prepare-host.yml
We moved the openstack_hosts setup into a separate task file so that the
openstack_hosts role could be tested individually, to avoid having to
call this specifically, we should include it in the
test-prepare-host.yml play as well.

This is required for Neutron OVS to work appropriately, and correctly
import the openvswitch module.

Change-Id: I521e80c00070f4338b84de3d61d9739645e00e12