Ensure ring always rebalances.
As this is for testing only, we should set the
"swift_pretend_min_part_hours_passed" variable to True. This will ensure
the ring is rebalanced and that we don't have to wait for an hour to
pass before any ring changes can be made.

This is especially important for MR swift where the ring will refuse to
rebalance on the second region's introduction, as min_part_hours has
not passed yet.

Change-Id: Ide83c1570e69e8fa2b0803e49215ef24d1fd2484