Add container logs to the build
Currently we don't have logs of what happened in the containers.

This commit changes the lxc_container_create default bind mounts
to ensure that the /var/log directory from the containers is
bind mounted to /var/log/{{ inventory_hostname }} on the host.

The tox configuration is also changed to ensure that links are
not dereferenced in the log copy, and that the layout of the log
storage is the same as that of the integrated gate.

Change-Id: I56f5e422313c405d4f65feb8e445d4f4b0700566
Co-Authored-By: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>