Allow storage and repl networks to be dynamic
For additional gates that may require swift (e.g. ironic), we can't
guarantee the networks that will be used for storage/replication - this
PR sets them as variables so that they can be overridden if required.

Additional fix to fix a regression that was fixed in the swift
repository before central tests were implemented.

Change-Id: Ia983acb3840358c9d0035ee3d69d6b4c4e4bfcf7