Default the variable if the group doesn't exist
We have situations where configuration files require host settings, but
they don't require the host_group. For example, nova requires ironic
variables but not ironic, or keystone with rabbitmq etc.

To prevent a situation where we require a host just to get the IP, or a
situation where we have to override the variable inside each role - lets
check if the group exists and set the variable accordingly.

Additionally, to prevent having to set the same long hostvar each time,
we can set that at the top of each "group" variables and then set those
variables appropriately.

Change-Id: I60d4079e913690069b5c44b83019a171dd3e4593