Allow network Alias addresses
For nova flat networking to work /w neutron we need to allow an alias
address for some interfaces.

This PR exposes a "alias:" option which is an address that will be added
as an alias for the interface.

For Debian this means an additional block at the bottom of the
configuration file.

For RedHat this means an <interface_name>:0 file is created.

Change-Id: I39c396ce3b6eab321d5eb70645c1d0af71dcf5c6