Remove workaround for bad pycparser wheel
This reverts https://review.openstack.org/381007 now that the
root cause of the issue [1] has been resolved, and that the
upper-constraints for OpenStack have been updated [2] to work
around the issue.

It also adds a clearer explanation for the reason why
libvirt-python is not installed via a wheel in tests.

[1] https://github.com/eliben/pycparser/issues/147
[2] https://review.openstack.org/381011

Change-Id: I2a4e75d952d4218fa7a0847e6fbf8f6afc15a65b