Fix neutron_agent IP to be neutron_server IP
The neutron_agent group was being used to determine the endpoint for
neutron server on port 9696. This worked because all neutron test
installs had the neutron_agent and neutron_server on the same IP.

The calico testing does not install neutron_agent however, as a result
this fails, this PR fixes that to point to the neutron_server (which is
the correct location for the endpoint).

Change-Id: I6f1f21cdbccc151b0305067085fe7645c6b0b504
(cherry picked from commit 22e02044d013d67637e32aa219633078421af920)