Prevent repeated get-pip.py fetches in role tests
In the integrated build, the get-pip.py script is put into the
repository, so the default of 'force: yes' is appropriate.

In role tests, however, we may wish to skip always fetching the
script to reduce errors relating to SSL validation when the task
executes a second time.

Depends-On: Ifb877e859a302b85c877aa867a0fc630b8765a19
Change-Id: I8b1c570c3a0713de3e84fb4e4068e270d9e5a4dd
(cherry picked from commit 4940cf1b24d7672961e7cd769f46624271f07516)