Add options for nova cell_v2
This patch adds the cell_v2 nova_cell0 database as part of the
test-install-nova.yml playbook. The var is required in test-vars for
this reason.

Additionally, we need to add permissions for the nova_api database user
to access the nova_cell0 database. This means we need to append
permissions for that user using the create-grant-db play. This defaults
to "omit" so is only used when "db_append_privs" is passed to the
include.

A work around is required for a bug in upstream nova - this should be
removed once https://review.openstack.org/#/c/428120/ merges.

Finally a fix was added to ensure the api_database is setup using the
appropriate address (although in this case it has no impact).

Change-Id: I082f37bb3ce61a900e06a58f21c7882f83671355