Temporarily bypass tests for Ubuntu Trusty
Ubuntu Trusty is not a supported platform for Ocata onwards.
Currently the CI system is executing jobs against Trusty
unintentionally and it's taking some time to unwind it.

This bypasses all tests if the platform detected is Trusty
so that development can continue while we work out the CI
configuration issue.

Change-Id: I236cf4aa748c22ad0249c1eb62e7fc64736f7f7f