Make tempest execute by default
Although the role executes to install it, tempest itself will
not execute unless told to do so. As a deployer may want
tempest to install but not execute in the playbook we set
the role default to not execute, so we must make it execute
here.

Change-Id: I269b43200c0e8773b5c0c437f685f6a0a4cd4e5e
(cherry picked from commit 52579f21feb16aa049de83741e1b827867b11c6a)