Enable cross-repo testing for OSA role tests
When 'Depends-On: <change-id>' is used in a patch,
Zuul cloner must be used in order to ensure that
both patches are actually used for the gate test.

This patch implements the use of zuul-cloner [1]
for gate tests and falls back to using git if
zuul-cloner is not present.

Additionally the playbook execution requires the
file based inventory to be used instead of the
CLI string so that the python interpreter is set.
This is required in order to ensure that the copy
module can act against CentOS with access to the
libselinux-python library on the host. See [2] for
more details.

[1] http://docs.openstack.org/infra/zuul/cloner.html
[2] https://goo.gl/VCKKfp

Change-Id: I8cf4d7d30284e5ef1e70897f2ecfe1da0b11c0c7