Add become to plays that delegate to localhost
'ansible_become' defined as a host_var or in the inventory for a host
has a higher precedence than 'become' defined within a playbook.

In change Id5b76a87809f03951c954fc3d752419a673403f7, 'become' was
defined explicitly for each play running against localhost. It will also
need to be defined in plays that delegate to localhost.

Change-Id: Ifb5e907594ca5fba459ee5a90b202e17db081f9a