Vagrantfile: Do not run the tox tests as root
By default, Vagrant runs the provision script as root. However, in the
OpenStack CI such tests are being executed as normal user so we observe
different behaviors when we compare the tests results between these two
platforms. As such, add 'privileged: false' to the Vagrantfile to
make Vagrant provision the VM as normal user.

Closes-Bug: 1682108
Change-Id: Ia4887d90d77110a88030a1958f57eddce5c4b2c2