Add become to all plays
The swift role includes tasks that delegate to the physical host of
containers (localhost in gate tests). Since ansible_become was removed
from localhost's vars, set become in this play to fix tests that install
swift.

Also add the become option to all other plays to avoid this issue for
any other roles that may now or in the future delegate tasks to localhost.

Change-Id: I08cada62636a1a7a1e2c1b15e0052876437e24aa