get-ansible-role-requirements: Do not clone the repositories as root
Commit cf6bea2acc06 ("Enable cross-repo testing for OSA role tests")
added the ability to test patchsets along with their dependencies.
However, it also modified the get-ansible-role-requirements.yml
playbook leading to repositories being cloned as root. This breaks
regular user testing as follows:

ln -s /vagrant /home/vagrant/.ansible/roles/<role>
ln: failed to create symbolic link ‘/home/vagrant/.ansible/roles/<role>’: Permission denied

The reason for that is that the $HOME/.ansible/roles is created as
'root' in the get-ansible-role-requirements.yml playbook so any attempt
to make modifications as regular user results to failure. It appears
that nothing in that playbook needs 'root' privileges so we run the
whole playbook as a regular user. Moreover, the playbook clears the
/etc/ansible/roles/* directories but the cloning happens in the $HOME
directory so we fix that as well.

Fixes: cf6bea2acc06 ("Enable cross-repo testing for OSA role tests")
Change-Id: I20c668108aea7404006247c9ba6abb1f5899735a