test-prepare-keys: Use ansible_user_gid for user's group
Using ansible_user_id to determine the primary group for a user implies
that such a group exists in the first place. Unfortunately this is not
always the case. For example, openSUSE does not create a new group for
every user but places all of them in the 'users' group. As such, avoid
using ansible_user_id and switch to ansible_user_gid which provides a
numeric value for the user's primary group.

Change-Id: I4fcea3e9bc9fa34998ab1a07e2c6bf60738e1fc1