SUSE: Remove post-up/post-down scripts from default ifcfg files
We do not need to use post-up/post-down scripts in the default ifcfg
files. These scripts are only installed when the ifcfg file handles a
veth pair. So even though we have them in the ifcfg files by default
nothing actually happens when we bring the interface up and down. This
can be very confusing to users who may wonder why we even have these
entries without installing the matching networking scripts.

The previous implementation was not quite correct anyway because
we were using item[0] and item[1] without having nested loops so the
results were sometimes unpredictable.

Change-Id: I14cbf062d67ad4f0e71c0fdeeaa89a4bdc12a4fe