sync-test-repos.sh: Add Zuul environment support for synchronizing files
We could re-use this script in the OpenStack CI propose job. For this to
happen we need to know if we are running in Zuul, and if we do, then we
need to consult the first argument which is normally the OSA project
we want to synchronize the files to. As such, add a new codepath which
only does a simple copy of files.

Change-Id: I7099fa2418b2633c378c37e83384fc78d7fbe554