Vagrantfile: Increase the ubuntu disk size to 40G
The official ubuntu-16.04 images are far too small to run tests for
complex roles such as the galera-server ones. This has been reported
upstream but no action was taken so far. As such, we need to increase
the disk size ourselves using the vagrant 'vagrant-disksize' plugin.
We increase it to 40G which is equal to what CentOS and openSUSE have.

Link: https://bugs.launchpad.net/cloud-images/+bug/1580596
Change-Id: Ifc0c62e243477ea5d83b2ce4a8edb3d08873c5f9