gen-projects-list.sh: Add ability to include extra OSA projects
Extend the script to allow us to define extra projects which should
be taken into consideration when syncing the test files. One such
project is 'ansible-hardening' which has replaced the
openstack-ansible-security one. This project doesn't match the usual
'openstack/openstack-*' pattern and as such it needs to be explicitely
added.

Change-Id: I97a5708b6897402f21ee87d5d637f7949059fd49