bindep.txt: Use explicit openSUSE Leap 42.2 dependencies
bindep.txt does not understand rpm patterns yet so we need to avoid
them until such support is added. As such, we use explicit package
names for openSUSE Leap 42.2 in order for experimental openSUSE CI
jobs to start being useful.

The bindep rpm pattern support is discussed in
https://review.openstack.org/#/c/476851/ so this patch can be
simplified once this change is in and deployed in the OpenStack CI
nodes. For now, this change is needed to unblock the openSUSE CI jobs.

Link: https://review.openstack.org/#/c/476851/
Change-Id: I1694d18c685b5bf2ced8517e3f8f5de97fcd5c2f