test-prepare-host: Fix POST_{UP,DOWN}_SCRIPT handling for SUSE
Commit 0a40e658cde96 ("SUSE: Remove post-up/post-down scripts from
default ifcfg files") dropped POST_{UP,DOWN}_SCRIPT from the ifcfg
file in the default case so they can only be used when a veth pair
is defined. However, the wrong destination was used in that case since
these statements were added in the script instead of the ifcfg file and
as a result they had no effect.

Fixes: 0a40e658cde96 ("SUSE: Remove post-up/post-down scripts from default ifcfg files")
Change-Id: I10b86b1bd1cd0a900e045bfc36b59e686888d69d