test-vars: Always build cryptography from source
cryptography may bundle openssl in the wheel and that causes symbol
conflicts if a different openssl is provided by the distribution.
As such, it's probably safer to re-build cryptography ourselves just
to be sure that the correct distro libraries are used. We may want to
revert that once we start building wheel packages for openSUSE and
distribute them in the OpenStack mirrors since every distribution
well then get a proper wheel file for its needs. See related review
https://review.openstack.org/#/c/486305/

Closes-Bug: 1705521
Link: https://github.com/pyca/cryptography/issues/3804
Change-Id: I7e88935acda580d8522a1e6927ea498431d78bda