test-log-collect.sh: Add SUSE/zypper handling when collecting logs
We fix the existing conditional to check for 'yum' instead of 'rpm' to
handle logs on RedHat and we also add an additional case to handle logs
on SUSE using the 'zypper' command. Moreover, we simplify the
conditional a bit to check for the exit code instead of the produced
output since if 'yum' or 'zypper' doesn't exist then we get a noisy
warning in the output. Fixes the following problem when collecting
logs on SUSE hosts:

2017-07-26 07:08:11.677110 | sudo: yum: command not found
2017-07-26 07:08:11.702615 | sudo: yum: command not found
2017-07-26 07:08:13.507998 | ERROR: InvocationError: '/bin/bash -c /home/jenkins/workspace/gate-openstack-ansible-os_keystone-ansible-func-opensuse-422-nv/tests/common/test-ansible-functional.sh'

Change-Id: I25c4ae9447726b7c72d35b8ee15f958d9fef507d