Use sudo with rsync for logs
Some logs in /var/log were skipped, such as yum.log, because rsync
is not being run as root. This patch uses sudo to rsync logs and
this should give us more of the logs we're looking for.

The yum.log will be very helpful to diagnose CentOS gate shenanigans.

Change-Id: I6abc69de66e15f336c22d216216be530165c5887