Install pyOpenSSL on CentOS
python2-pyOpenSSL is only available from the CentOS RDO QEMU repo
and that repo is not commonly installed on a basic CentOS 7 system.
It happens to be pre-installed in the OpenStack CI gates and that's
why this bindep file works.

However, this package installation fails on a normal CentOS 7
system and it can cause problems during deployment or testing.

The EPEL repository has a package called pyOpenSSL that provides
python bindings for OpenSSL on CentOS systems. That should be used
instead.

In addition, dnf will fail entirely with a bad exit code if a package
isn't found. The yum package manager will trudge on without errors.

Closes-Bug: 1708493
Depends-On: I2ac5a9c463991e4668a3e7608831c1944377e85e
Change-Id: I93de939304b57a211f2eb281365d5a591752be7d