Use LXC reverse proxy and improve UCA repo discovery
This patch implements the use of the images.linuxcontainers.org
reverse proxy when in nodepool, and also improves the discovery
of the UCA repo URL so that it's done in the way that infra
expects it to be done.

The lxc_container_template_main_apt_repo and
lxc_container_template_security_apt_rep fact overrides are
removed as they serve no purpose any more. Those vars were
removed in Newton.

Depends-On: Ia6b149f157e5697d36fdea3dc65cb2ac4815a1a4
Change-Id: I0e27abaea4cac010e9ecdae4dcb1d6cb83ad9d29