test-ansible-functional.sh: Only run yum-config-manager on CentOS
The existing conditional was wrong since it matched the Fedora
distribution which doesn't need to run yum-config-manager. As such,
we only run it if it's available and if it's a CentOS host.

This is similar to what is being proposed for the project-config
repository at https://review.openstack.org/#/c/493643
Change-Id: I0f321b9d5a6df64a9a0bc93c53738fa9a20b63dc