Implement permission changes during rsync
In order to ensure that the files and directories
that are gathered for logs are readable for the
rename and gzip processes, we ensure that the
permissions and ownership are set during the copy
handled by rsync.

The previous method resulted in the files being
copied, but some files not being writable when
executing mv/gzip. This was noticed in the
os_neutron role where the calico job was failing
due to permissions issues for the felix config
files in the container /etc directory.

Change-Id: I7b84e0db25f44842a88315cd413ac39f75f9f8b9
(cherry picked from commit 0426d7f51b5bea9f7dc2baf3c3401565d016ba47)