Only report errors when collecting logs
Currently there are a lot of log collection lines
showing which files are not copied because they
are symlinks. They take up an awful lot of room
in the console log and are not very useful.

This patch sets the --quiet flag so that only
errors show up in the console log.

Change-Id: I2e03bf3f0a5b1aff3c4cc6a4e4803d172dabde65