Log begin/end of log collection
The console log has output from the log collection,
but it's not clear that the output is part of a test
or something after the test.

This makes it clearer that the output between the
begin/end are related to log collection.

Change-Id: Iba374d89fa5c11f8690af73e87804cb29cb7fc9e
(cherry picked from commit 4bda4b9ea0bb4db63604d14fb44316b32c3f27b3)