Split log permission changes from rsync command
Unfortunately the version of rsync used by CentOS
(3.0.x) does not support the use of the --chown
option, so we need to go back to doing it in a
separate command.

The rsync command is adjusted to ignore ownership
and permissions when copying, reverting to the
target folder defaults. Then in the log folder
we make sure it has permissions for the jenkins
user. The chmod is done first to ensure that all
dirs are traversable.

Some line breaks are added to the script to help
make it a little more readable.

Change-Id: I74157f7bc449ff7ee6d2e280dff54b2c9ebfa9f0