Unwedge the openstack-ansible-tests gate
This is a combined patch that does two things:

1. Set the LXC download mirror properly

The LXC reverse proxy from nodepool was not being used by the
lxc_hosts role and this patch fixes it.

Closes-Bug: 1716958

2. Pin the keystone role SHA

Keystone has removed various v2 APIs but tempest still wants to
test them. This buys us some time to get tempest fixed and
we can remove the pin.

Related-Bug: 1717011
Change-Id: Idf55daa853ff73602bfe6d2862080bfd874e17f4