Move log collection to 'post' stage
To ensure that logs are collected regardless of whether
a job succeeds, fails or times out, we move the log
collection to the 'post' stage of the job definition.

As the job playbook only gets run in nodepool, we remove
the detection of nodepool in the log collection script
and also remove the trap which executes it so that it
is only ever run when executed in zuul.

Change-Id: I5ab31cc141804cbcb6d8721032a79d0c724fc0a2