Add commonly used repositories to required projects
Zuul is able to more effectively clone the roles used in
role tests more reliably that using a standard git clone
because it has a cache implementation to do exactly that.

This implements the commonly used role repositories as
required-projects for the tests so that our playbook to
get the ansible-role-requirements deoesn't have to do
the cloning, it'll just do the right symlinking. It will
still fetch any missing repositories, so this list does
not have to be a *full* list - just enough to help reduce
the current failures due to git cloning problems.

Change-Id: I203dce5ea87feed7c4354039f035742c9c3a5f36