Use new zuul v3 git src location
Previously we used a complex job definition to make sure
that the appropriate git clones for cross-repo testing
were in place. Now we just used the same location where
zuul v3 puts the git sources automatically when a
cross-repo test is done.

This ensures that our cross-repo tests do what they're
supposed to and that we're using the same default location
as the rest of openstack's tests, making it more intuitive.

Change-Id: Ia5ab6aa43f912b3cb4adf98154cd3da12688c2e0