Ensure all pre-cloned git sources are found
Currently the find task only looks at one level deep
for repo folders, instead of the required three levels
deep.

Using the find module does not give us the ability to
set the depth level we want to do the find down to,
so we use a find command instead.

Change-Id: I0b218fe276292e13dbb40ab38c9a167025d16e45