Consolidate run & post playbooks for all tests
When zuul executes jobs it adds all the post plays together
and executes them in the inherited sequence. This is causing
some confusion in our cross-repo testing where the logs are
trying to be collected twice as the cross-repo test is based
on the standard functional test.

Instead of defining two sets of plays we can consolidate them
into a single set, eliminating the confusion and repetition.

At the same time we also do not try to collect logs if the
script to prepare them was not able to execute, and we fix
the conditional for the ARA report generation so that it
works properly with the value being given it by zuul which
is a boolean word. To make sure it still works outside of
Zuul we leave the conditional which uses the standard success
return code of 0.

Change-Id: Id5027415f33a7a41650332a983cb87fab7ba2a4a