Implement retries for git clones
There have been several failures due to git clone TLS handshake
fails.

This patch implement the use of clone depth in order to reduce
the amount of data being cloned. This will not always work as
depth only works when the clone is from a ref, but it will help
in those circumstances.

This patch also implements retries with a delay between each
retry in order to try and improve the success rate.

Change-Id: I72a7d39218c307fc1c6f108b81f35d2cd5bb5c66