Force SUSE repo refresh and switch default mirror.
This change forces a repo refresh whenever running tests on SUSE. This
is needed to ensure all of the repos are up-to-date and we're pulling
package from the correct sources. This also sets the default mirror
to download.opensuse.org since the OpenStack mirrors are outdated due
to rsync.opensuse.org being broken. See https://progress.opensuse.org/issues/27699

Change-Id: I9f2a1f98981cdb444c2162b09c02f6e530f5d1c2
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
Signed-off-by: Markos Chandras <mchandras@suse.de>