Remove the serialization in tests
The lxc container create process were running in serial as it sets up
the test environment. Being this is setting up a test environment
there's no need for it to be executed in serial. The original reason
these container creates were running in serial was due to SSH issues
however the integrated repo has been creating containers in parallel for
sometime now so it is believed the original control path issues have
been resolved.

Change-Id: I59d563549b288e634338474aed45011e94b85c47
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>