Always generate ARA subunit results
We'd like to see our tests results show up in OpenStack Health
regardless of whether the test passed or failed. This patch
makes sure that it will always try to generate them.

It also implements a routine to try and find where ARA is
installed, because it's not always in a tox venv called
'functional' which was the previously hard-coded location.

Change-Id: I3d967567af9d9c9583f34c89789a38e60717781e
(cherry picked from commit 2fb00b67edb249d47a422434757f406926b747a0)