Vagrantfile: Use rsync to copy working directory to VM
Depending on the VM distro the working directory can either be rsync'd
to the VM or mounted. When the directory is mounted, various artifacts
are left behind from the test execution which is not desired especially
when the directory is shared between multiple VMs for parallel testing.
As such, it's best to simply copy the directory to the VM using 'rsync'
so each VM can get its own copy. This also restores the missing new line
at the end of the test-log-collect.sh script which got removed in
I4475e4dfe13acb4c88444f1f7aa88272bf87f6ee in order to make the linters
happy.

Change-Id: I1aab7d3f0d26f35dea84ae088fe78cfd2de79366