Use a venv with previous ansible version for upgrades
When doing the role upgrade, we run every playbook with current's
branch ansible. This is normal for the infrastructure bits (which
are running on current branch), but not for the installation of
the previous branch's role, which should use previous branch's
environment. This script already exists for some OSA roles to
facilitate upgrades but we are now moving it to the tests
repository so all roles can benefit from it.

(Original patch 91b7524603650edbef984fd9fe83e352b90604d2)
Link: https://github.com/ansible/ansible/pull/34790
Co-Authored-By: Jean-Philippe Evrard <jean-philippe@evrard.me>
Closes-Bug: #1741471

Change-Id: I7ebb045885dd645b820de2b7f75b46c755c015f6
(cherry picked from commit f51a22988de85e44b122af8046eedc396de98032)