Add memcached servers wiring
By default, we could make sure keystone is used with memcached.
For that we wire the new directive keystone_cache_servers to
the memcached hosts from the group that's defined in
test_memcached_group.

If the role has no memcached_all group, it would use the
all_containers group, while if it has it, it would use it
directly. That allows testing multiple scenarios if need be.

Change-Id: I1be8ad46c29c1719c7fadd4c91aeea21923d5681
Depends-On: I85ebce8b41dd440e1866a08aa1329b3df798c04f