Update rabbitmq install for cluster and upgrade tests
The current rabbitmq installation playbook assumes that there is
only a single rabbitmq container. While this works for most roles,
when more than one is desired in a test it requires a duplicated
playbook in the role which behaves differently.

This patch updates the playbook to make it more universal.

Change-Id: I372e0db39989ef412ab4301fcd5493efb48bdbe9