Set a fact for handling yum's fastestmirror plugin
In the past, we've messed around with enabling or disabling the
yum's fastestmirror plugin. It is helpful if you are running outside
the OpenStack CI system since it will find the fastest mirrors from
a publicly available mirror list.

However, in the CI system, we override the mirror list with a hard-
coded mirror in the `baseurl` line. There's no need to load or use
the fastestmirror plugin in these situations and gathering a list
of mirrors just slows things down.

This patch sets a fact that other roles can check and disable the
fastestmirror plugin when needed.

Change-Id: I8b2a1db084ab47962dbef8c6b037836f44b4ed46